requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.9.0
python-multipart==0.0.12
python-dotenv==1.0.1
requests==2.32.3
groq==1.0.0
mangum==0.17.0
scikit-learn==1.4.2
joblib==1.3.2
